Add static casts

+++ wReceiver-base.cpp
@@ -69,7 +69,7 @@ int main(int argc, char *argv[]) {
 
         // Only accept START if no connection in progress or sender retries
         if (connection_seq_num == -1 ||
-            connection_seq_num == packet.header.seqNum) {
+            connection_seq_num == static_cast<int>(packet.header.seqNum)) {
           // Send acknowledgement
           send_ack(server_addr, sockfd, log, packet.header.seqNum);
 
@@ -77,7 +77,7 @@ int main(int argc, char *argv[]) {
         }
 
       } else if (packet.header.type == DATA && connection_seq_num != -1 &&
-                 packet.header.seqNum < window_end && valid_checksum(packet)) {
+                 static_cast<int>(packet.header.seqNum) < window_end && valid_checksum(packet)) {
         // Mark packet as received
         packets_received[packet.header.seqNum] = packet;
 
@@ -91,7 +91,7 @@ int main(int argc, char *argv[]) {
         // If the current connection is closing
         std::string output_file =
             "FILE-" + std::to_string(connection_count) + ".out";
-        if (connection_seq_num == packet.header.seqNum) {
+        if (connection_seq_num == static_cast<int>(packet.header.seqNum)) {
           std::string output_path = std::filesystem::current_path() /
                                     output_dir.substr(1) / output_file;
           std::cout << "current_path " << std::filesystem::current_path()
